NET-85: bootstrap and authorization sequence changed
