+++ network/controller/bootstrap/helper_test.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package auth
+package bootstrap
 
 import (
 	"testing"
